fix a bunch of small bugs

+++ Client/Game.h
@@ -28,7 +28,7 @@ struct rr_game_loadout_petal
 struct rr_game_squad_client
 {
     struct rr_game_loadout_petal loadout[20];
-    float requested_start_wave;
+    uint32_t requested_start_wave;
     char name[18];
     uint8_t in_use;
     uint8_t ready;
